Add ros2 messages, interfaces and update to colcon build system

+++ ocs2_core/include/ocs2_core/control/LinearController.h
@@ -97,13 +97,13 @@ class LinearController final : public ControllerBase {
 
   scalar_array_t controllerEventTimes() const override;
 
-  void flatten(const scalar_array_t& timeArray, const std::vector<std::vector<float>*>& flatArray2) const override;
+  void flatten(const scalar_array_t& timeArray, const std::vector<std::vector<double>*>& flatArray2) const override;
 
   static LinearController unFlatten(const size_array_t& stateDim, const size_array_t& inputDim, const scalar_array_t& timeArray,
-                                    const std::vector<std::vector<float> const*>& flatArray2);
+                                    const std::vector<std::vector<double> const*>& flatArray2);
 
  private:
-  void flattenSingle(scalar_t time, std::vector<float>& flatArray) const;
+  void flattenSingle(scalar_t time, std::vector<double>& flatArray) const;
 
  public:
   scalar_array_t timeStamp_;
